Keep sourceDirs and generatedSourceDirs mutable by adding to existing… (#625)
Keep sourceDirs and generatedSourceDirs mutable by adding to existing set.

Instead of setting sourceDirs and generatedSourceDirs to be an ImmutableSet, retrieve the existing respective sets and add to them. This is the commonly accepted method of appending sourceDirs and generatedSourceDirs, and keeps each set mutable to avoid breaking other plugins.